+++ src/HelloWorld.js
@@ -1,5 +1,5 @@
 "use strict";
-// Example A: HelloWorld.ts
+// Example: HelloWorld.ts
 var __extends = (this && this.__extends) || (function () {
     var extendStatics = function (d, b) {
         extendStatics = Object.setPrototypeOf ||