+++ src/RedirectHttps.js
@@ -1,5 +1,5 @@
 "use strict";
-// Example C: RedirectHttps.ts
+// Example: RedirectHttps.ts
 var __extends = (this && this.__extends) || (function () {
     var extendStatics = function (d, b) {
         extendStatics = Object.setPrototypeOf ||